Fully integrated linear operators into lattice.so and lattice.py
Changed exposed propagator function to be wilson-specific. Other fermion
actions should be exposed using different functions. It is now also
possible to specify the type of smearing to be performed (thought there
is still only one type available), as well as the solver parameters
(tolerance and maximum number of iterations).

These changes come at the expense of the ability to use CUDA, which
still needs to be implemented using the linear operator technique that
the CPU code now uses.