Added cuda kernels and linear operators.
The added files contain code that resembles that src/linear_operators.
Linear operators are based on the cusp::linear_operator type. I've
created a LinearOperator base class, just like in src/linear_operators,
which transfers the links to the GPU. Everything is based of this class.

The CUDA kernels are in kernels.h and kernels.tcu, and some utilities
are in utils.h and utils.tcu. There's a generalised hopping kernel,
along with a diagonal kernel.

Since utils and kernels use templates (in the hope of optimizing the
calculation), the implementation for these is in tcu files, which are
included at the end of the header files.