Moved file input/output functions into io submodule
The existing xml parsing and data archiving functions are all fileio
functions, so it makes sense to put them in their own io submodule. I've
also added functions to convert various binary config files to numpy
arrays, using qcdutils to do so. Some changes were needed in qcdutils to
make it compatible with Python 3, and I've stripped out various
functions that are used by the main function but are otherwise not
needed.