Refactored LatticeBase a bit
Moved accessor functions required by expression templates to private,
and made expression template classes friends of LatticeBase to allow
them access. Then added interface class in tests to allow access to
these private member variables during tests. This hides the inner
workings of the class from the user.

Made destructor implementation of the class part of the declaration, as
nothing happens there and it was getting in the way.