Fairly major refactorization of LatticeBase expression templates
To reduce code repetition, all binary expression templates are now
replace with the LatticeBaseBinary template. There is now also a
LatticeBastConst class to use with the new binary template when scalar
multiplication or division is required. I've also added a traits system
for this class to ensure that only references to the LatticeBase class
are stored, to prevent any accidental copies, and to correctly store any
temporary LatticeBaseConst objects.