Gone through yet another complete revamp
Right, so after deciding using pure Cython for the majority of the code,
I realised this was just totally unworkable, as the performance loss
would be significant. So, with that in mind, the plan now is to utilise
numpy for the front-end as much as possible, but write the core actions
(gauge and fermionic) in C++, as well as the update algorithms etc.
Hopefully this way everything will be more performant.