Removed Complex and Layout types, plus further tidying to Cython codegen.
Exposing the Layout class to Python is just going to make things more
complicated on the NumPy side, so it's been removed. Instead, lattice
types now accept the shape of the lattice in the constructor argument.
The Complex type isn't actually referred to anywhere anymore, so that's
been removed too. Finally, I've moved some of the template code
generation stuff inside the ContainerDef class, to make it easier to
read.